Extract `useDateFilter` hook
