+++ frontend/src/metabase/common/hooks/filters/use-date-filter/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-date-filter";
